exposure model view template
